refactor(generate-g2): extract args and model dto modules
