+++ crates/assay-cli/src/cli/commands/generate.rs
@@ -11,8 +11,6 @@
 //! ```
 
 use anyhow::Result;
-use clap::Args;
-use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use std::io::{BufRead, BufReader};
 use std::path::{Path, PathBuf};
@@ -21,161 +19,15 @@ use super::events::Event;
 use super::heuristics::{self, HeuristicsConfig};
 use super::profile_types::{self, stability_smoothed, Profile, ProfileEntry};
 
+mod args;
+mod model;
+
+pub use args::GenerateArgs;
+pub use model::{serialize, Entry, Meta, NetSection, Policy, Section};
+
 // ─────────────────────────────────────────────────────────────────────────────
 // CLI Args
 // ─────────────────────────────────────────────────────────────────────────────
-
-#[derive(Args, Debug, Clone)]
-#[command(about = "Generate policy from trace or profile")]
-pub struct GenerateArgs {
-    /// Input trace file (single-run mode)
-    #[arg(short, long)]
-    pub input: Option<PathBuf>,
-
-    /// Profile file (multi-run mode)
-    #[arg(long)]
-    pub profile: Option<PathBuf>,
-
-    #[arg(short, long, default_value = "policy.yaml")]
-    pub output: PathBuf,
-
-    #[arg(long, default_value = "Generated Policy")]
-    pub name: String,
-
-    #[arg(long, default_value = "yaml")]
-    pub format: String,
-
-    #[arg(long)]
-    pub dry_run: bool,
-
-    /// Show policy diff versus existing output file
-    #[arg(long)]
-    pub diff: bool,
-
-    // ─── Single-run heuristics ───
-    #[arg(long)]
-    pub heuristics: bool,
-
-    #[arg(long, default_value_t = 3.8)]
-    pub entropy_threshold: f64,
-
-    // ─── Profile stability ───
-    /// Minimum stability to auto-allow (profile mode)
-    #[arg(long, default_value_t = 0.7)]
-    pub min_stability: f64,
-
-    /// Below this, mark as needs_review if --new-is-risky
-    #[arg(long, default_value_t = 0.6)]
-    pub review_threshold: f64,
-
-    /// Treat low-stability items as risky (else skip them)
-    #[arg(long)]
-    pub new_is_risky: bool,
-
-    /// Smoothing parameter (Laplace) for display
-    #[arg(long, default_value_t = 1.0)]
-    pub alpha: f64,
-
-    /// Minimum runs before anything can be auto-allowed (safety belt)
-    #[arg(long, default_value_t = 1)]
-    pub min_runs: u32,
-
-    /// Z-score for Wilson lower bound gating (1.96 ≈ 95% confidence)
-    #[arg(long, default_value_t = 1.96)]
-    pub wilson_z: f64,
-}
-
-impl GenerateArgs {
-    pub fn validate(&self) -> Result<()> {
-        if self.min_stability < 0.0 || self.min_stability > 1.0 {
-            anyhow::bail!("--min-stability must be between 0.0 and 1.0");
-        }
-        if self.review_threshold < 0.0 || self.review_threshold > 1.0 {
-            anyhow::bail!("--review-threshold must be between 0.0 and 1.0");
-        }
-        if self.min_stability < self.review_threshold {
-            anyhow::bail!(
-                "--min-stability ({}) must be >= --review-threshold ({})",
-                self.min_stability,
-                self.review_threshold
-            );
-        }
-        if self.alpha <= 0.0 {
-            anyhow::bail!("--alpha must be positive");
-        }
-        if self.wilson_z <= 0.0 {
-            anyhow::bail!("--wilson-z must be positive");
-        }
-        if self.entropy_threshold < 0.0 {
-            anyhow::bail!("--entropy-threshold must be non-negative");
-        }
-        Ok(())
-    }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Output Types
-// ─────────────────────────────────────────────────────────────────────────────
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default)]
-pub struct Policy {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub _meta: Option<Meta>,
-    pub files: Section,
-    pub network: NetSection,
-    pub processes: Section,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default)]
-pub struct Meta {
-    pub name: String,
-    pub generated_at: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub profile_runs: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_stability: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_runs: Option<u32>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default)]
-pub struct Section {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub allow: Vec<Entry>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub needs_review: Vec<Entry>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub deny: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Default)]
-pub struct NetSection {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub allow_destinations: Vec<Entry>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub needs_review: Vec<Entry>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub deny_destinations: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(untagged)]
-pub enum Entry {
-    Simple(String),
-    WithMeta {
-        pattern: String,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        count: Option<u32>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        stability: Option<f64>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        runs_seen: Option<u32>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        risk: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        reasons: Option<Vec<String>>,
-    },
-}
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Single-Run Aggregation
@@ -860,17 +712,6 @@ fn print_policy_diff(diff: &PolicyDiff, output_path: &Path) {
         "  Summary: +{} added, -{} removed, ~{} changed",
         added, removed, changed
     );
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Serialization
-// ─────────────────────────────────────────────────────────────────────────────
-
-pub fn serialize(policy: &Policy, format: &str) -> Result<String> {
-    Ok(match format {
-        "json" => serde_json::to_string_pretty(policy)?,
-        _ => serde_yaml::to_string(policy)?,
-    })
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
